moved positioning on components
