+++ ui/src/components/Dashboard.js
@@ -11,7 +11,7 @@ import Paper from '@material-ui/core/Paper';
 const styles = theme => ({
     root: {
         width: '100%',
-        marginTop: theme.spacing.unit * 5,
+        marginTop: theme.spacing.unit * 3,
         overflowX: 'auto',
     },
 });
