+++ ui/src/components/ReadWriteCard.js
@@ -9,7 +9,7 @@ import Grid from '@material-ui/core/Grid';
 const styles = theme => ({
     root: {
         marginTop: '10px',
-        marginLeft: '30px',
+        marginLeft: '100px',
         width: '95%',
     },
     cardFont: {
@@ -26,17 +26,17 @@ class ReadWriteCard extends React.Component {
         return (
             <Grid container className={classes.root}>
                 <Grid item xs={6}>
-                    <Card style={{width: "45rem", marginBottom: "12px"}}>
-                        <CardHeader color="primary" className={classes.cardFont}>PURCHASES</CardHeader>
-                        <CardBody style={{height: "80px"}}>
+                    <Card style={{width: "37rem", marginBottom: "12px"}}>
+                        <CardHeader color="primary" className={classes.cardFont}>"PURCHASES"</CardHeader>
+                        <CardBody style={{height: "120px"}}>
                             <h4>...purchase data will load here</h4>
                         </CardBody>
                     </Card>
                 </Grid>
                 <Grid item xs={6}>
-                    <Card style={{width: "45rem", marginBottom: "12px"}}>
+                    <Card style={{width: "37rem", marginBottom: "12px"}}>
                         <CardHeader color="warning" className={classes.cardFont}>EVENT LOG</CardHeader>
-                        <CardBody style={{height: "80px"}}>
+                        <CardBody style={{height: "120px"}}>
                             <h4>...event logging data will load here</h4>
                         </CardBody>
                     </Card>
